% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal_fluency}
\alias{animal_fluency}
\title{Animal fluency data.}
\format{A list of 201 animal fluency sequences named by the participants age.}
\source{
data obtained from published papers. See referneces.
}
\usage{
animal_fluency
}
\description{
A dataset containing the animal fluency sequences of a lifespan sample of 201
individuals collected at Stanford university. See references.
}
\references{
Hills, T. T., Mata, R., Wilke, A., & Samanez-Larkin, G. R. (2013). Mechanisms of age-related decline in memory search across the adult life span. Developmental psychology, 49(12), 2396.
Wulff, D. U., Hills, T. T., Lachman, M., & Mata, R. (2016). The Aging Lexicon: Differences in the Semantic Networks of Younger and Older Adults. Cognitive Science Society.
}
\keyword{datasets}
