% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{cc}
\alias{cc}
\title{Local clustering coefficient}
\usage{
cc(g, type = "corrected")
}
\arguments{
\item{g}{a graph object of class \code{igraph}.}
}
\value{
the corrected local clustering coefficient and/or
the uncorrected clustering coefficient.
}
\description{
Computes the corrected average local clustering coffiecient
using \code{igraph}'s average \link[igraph]{transitivity}.
}
\details{
Per default the corrected clustering coefficient ist computed
which is computed as the average transitivity divided
by the \code{average degree / n vertices}, i.e., the expected
average local transitivity for an Erdös-Renyi random graph.
}
