% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{common_subgraphs}
\alias{common_subgraphs}
\title{Get common subgraph}
\usage{
common_subgraphs(adj_1, adj_2, weights_1 = NULL, weights_2 = NULL,
  mode_1 = "undirected", mode_2 = "undirected")
}
\arguments{
\item{adj_1}{numeric matrix representing the adjacency matrix of graph 1.}

\item{adj_2}{numeric matrix representing the adjacency matrix of graph 2.}

\item{weights_1}{numeric vector of edge weights for network 1 and
2, respectively. Optional.}

\item{weights_2}{numeric vector of edge weights for network 1 and
2, respectively. Optional.}

\item{mode_1}{character, either \code{"directed"} or \code{"undirected"},
specifying whether network 1 and 2 should be interepeted as directed
or undirected, respectively. Defaults to \code{"undirected"}.}

\item{mode_2}{character, either \code{"directed"} or \code{"undirected"},
specifying whether network 1 and 2 should be interepeted as directed
or undirected, respectively. Defaults to \code{"undirected"}.}
}
\value{
List containing the two subgraphs.
}
\description{
Function identifies and returns the common subgraphs of two networks.
}
