% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fluency_steps}
\alias{fluency_steps}
\title{Verbal fluency step counter}
\usage{
fluency_steps(adjlist, n, pjump = 0, type = 0L)
}
\arguments{
\item{n}{integer vector specifying the numbers of production.}

\item{pjump}{numeric specifying the probability of a jump.}

\item{type}{integer controlling network start and jump nodes.
For \code{type = 0} the process selects the start node and any jump
nodes proportional to their degree. For \code{type = 1} the process
selects a random node to serve both as the start node and the jump node.
For \code{type = 2} the process selects the start and any jump nodes
uniformly at random.}
}
\value{
List of character vectors containing the indices of the fluency productions.
  Indices refer to the row of the item in the original adjacency matrix. See
  \link{get_adjlist}.
}
\description{
Repeatedly generates verbal fluency data using \link{one_fluency_steps} and
counts the number of steps required to produce \code{n} unique responses.
}
\details{
For details see \link{one_fluency_steps}.
}
