% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{g_stat}
\alias{g_stat}
\title{Graph statistics}
\usage{
g_stat(g, giant = FALSE, directed = F)
}
\arguments{
\item{g}{a graph object of class \code{igraph}.}

\item{directed}{logical specifying whether the graph should
be interpreted as being directed. Applies for \code{aspl}
and \code{assortivity}.}

\item{gian}{logical specifying whether the largest component
of \code{g} should be used instead of \code{g}.}
}
\description{
Function computes various graph statistical measures
}
