% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_kneighbor}
\alias{get_kneighbor}
\title{Get vector of neighbors exactly k steps away}
\usage{
get_kneighbor(adj, start, k)
}
\arguments{
\item{adj}{numeric matrix specifying the adjacency matrix.}

\item{start}{integer specifying the row of the start vertex in the adjacency matrix.}

\item{k}{integer specifying the exact distance to the start vertex.}
}
\value{
A character vector containing vertices \code{k} or fewer steps away
from \code{start}.
}
\description{
Function iterates over graph to identify for a given vertex all vertices that are
exactly k steps apart.
}
\details{
k = 0 will return be interpreted as k = 1.
}
