% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_names}
\alias{get_names}
\title{Get node names of memnet objects}
\usage{
get_names(dat)
}
\arguments{
\item{dat}{numeric or character matrix containing an edgelist or adjacency
matrix.}
}
\value{
lists of character vectors.
}
\description{
Function extracts names from objects created from edgelists and adjacency
matrices.
}
\examples{

# get fluency data
data(animal_fluency)

# edge list of fluency graph
edge_list = threshold_graph(animal_fluency)

# get names
get_names(edge_list)

}
