% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{goni_graph}
\alias{goni_graph}
\title{Create Goni graph}
\usage{
goni_graph(dat, l = 3L, min_cooc = 1L, crit = 0.05)
}
\arguments{
\item{l}{an integer specifying the window size. The internal upper limit
of \code{l} is the number of productions.}

\item{min_cooc}{integer specifying the minimum number of times two words
have to coocur within a window size of \code{l} to consider including
an edge between them.}

\item{crit}{a numeric within \code{[0,1]} specifiying the type-1 error
rate of including an edge between unconnected words.}
}
\value{
A matrix
}
