% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grplot.R
\name{graph_plot}
\alias{graph_plot}
\title{Plot graph}
\usage{
graph_plot(g, vnames = NULL, layout = layout.fruchterman.reingold,
  grd_res = 48, horiz_lim = 3, vert_lim = 3, ver_col = "steelblue",
  edg_col = "grey50", txt_col = "black", shadow = T, pt.cex = 2.5,
  cex = 1, lwd = 2)
}
\arguments{
\item{g}{graph object of class \code{igraph}. Alternatively, a
matrix or data.frame with the first two columns specifying the
from- and to-vertices.}
}
\description{
Custom graph plot using \code{igraph}'s layout functions.
}
