% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grow_lattice}
\alias{grow_lattice}
\title{Regular lattice network model}
\usage{
grow_lattice(n = 100L, k = 10L)
}
\arguments{
\item{n}{Integer. Number of nodes in the network.}

\item{m}{Integer. Number of edges added for each incoming node.}
}
\value{
n x n adjacency matrix.
}
\description{
Grow regular lattice networks, in which every node is connected to m neighbors.
}
