% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grow_st}
\alias{grow_st}
\title{Steyvers and Tenenbaum (2004) network growth model}
\usage{
grow_st(n = 100L, m = 5L)
}
\arguments{
\item{n}{Integer. Number of nodes in the network.}

\item{m}{Integer. Number of edges added for each incoming node.}
}
\value{
n x n adjacency matrix.
}
\description{
Grow networks using Steyvers and Tenenbaum (2004) model, which combines
preferential attachment with a triad formation.
}
