% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacknife.R
\name{jacknife}
\alias{jacknife}
\title{Match group fluency data}
\usage{
jacknife(g, measures = c("k", "cc", "aspl", "smallw", "assor"),
  lcomp = TRUE)
}
\arguments{
\item{data}{a data frame containing (at least) fluency productions,
subject id, and a grouping variable.}

\item{labels}{a character vector containing the labels of variables
containing the fluency productions, subject id, and grouping.}
}
\value{
Function returns two lists with matching numbers of fluency
productions.
}
\description{
Prune fluency data of two groups so that the groups'
distributions of number of productions match.
}
