% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mfluency}
\alias{mfluency}
\title{Verbal fluency generator wrapper}
\usage{
mfluency(adjlist, n, pjump = 0, type = 1L)
}
\arguments{
\item{adjlist}{a list containing row indices for adjacent vertices as created
by \link{get_adjlist}.}

\item{n}{integer specifying the number of productions.}

\item{pjump}{numeric specifying the probability of a jump.}

\item{type}{integer controlling network start and jump vertices.
For \code{type = 0} the process selects the start vertex and any jump
vertices proportional to their degree. For \code{type = 1} the process
selects a random vertex to serve as the start vertex and the jump vertex.
For \code{type = 2} the process selects the start and any jump vertices
uniformly.}
}
\value{
List of character vectors containing the indices of the fluency productions.
  Indices refer to the row of the item in the original adjacency matrix. See
  \link{get_adjlist}.
}
\description{
Generates multiple verbal fluency sequences using \code{fluency}.
}
\details{
For details see \link{fluency}.
}
