% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mfsearch}
\alias{mfsearch}
\title{Fast verbal fluency generator wrapper}
\usage{
mfsearch(adjlist, n, pjump = 0, random = FALSE)
}
\arguments{
\item{adjlist}{a list containing row indices for adjacent vertices as created
by \link{get_adjlist}.}

\item{n}{integer specifying the number of productions.}

\item{pjump}{numeric specifying the probability of a jump.}

\item{random}{bool controlling jump vertices.
For \code{random = TRUE} the process selects jump at random. For
\code{random = FALSE} the process always jumps back to the start vertex. The
start vertex is always selected at random.}
}
\value{
List of character vectors containing the indices of the fluency productions.
  Indices refer to the row of the item in the original adjacency matrix. See
  \link{get_adjlist}.
}
\description{
Generates multiple verbal fluency sequences using \code{fluency}.
}
\details{
For details see \link{ffluency}.
}
