% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{one_fluency_steps}
\alias{one_fluency_steps}
\title{Verbal fluency step counter}
\usage{
one_fluency_steps(adj_list, n, pjump, type)
}
\arguments{
\item{n}{integer specifying the number of productions.}

\item{pjump}{numeric specifying the probability of a jump.}

\item{type}{integer controlling network start and jump nodes.
For \code{type = 0} the process selects the start node and any jump
nodes proportional to their degree. For \code{type = 1} the process
selects a random node to serve both as the start node and the jump node.
For \code{type = 2} the process selects the start and any jump nodes
uniformly at random.}

\item{random}{bool controlling jump nodes.
For \code{random = TRUE} the process selects jump at random. For
\code{random = FALSE} the process always jumps back to the start node. The
start node is always selected at random.}
}
\value{
Integer vector containing the indices of the fluency productions.
  Indices refer to the row of the item in the original adjacency matrix. See
  \link{get_adjlist}.
}
\description{
Generates verbal fluency data using a switcher-random walk process and counts
the number of steps required to produce \code{n} unique responses.
}
