% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{one_search}
\alias{one_search}
\title{Search network using switcher-random walk process}
\usage{
one_search(adj_list, start, observe, nmax = 1000L, pjump = 0,
  type = 0L)
}
\arguments{
\item{start}{integer vector of length 1 specifying the index of
the start node.}

\item{observe}{integer vector of length 1 or larger specifying the target
end nodes.}

\item{nmax}{integer specifying the maximum number of steps before search
terminates.}

\item{pjump}{numeric specifying the probability of a jump.}

\item{type}{integer controlling network start and jump nodes.
For \code{type = 0} the process selects the start node and any jump
nodes proportional to their degree. For \code{type = 1} the process
selects a random node to serve both as the start node and the jump node.
For \code{type = 2} the process selects the start and any jump nodes
uniformly at random.}
}
\value{
Numeric, 3 column matrix containing in each row the start node, the
end node, and the (minimum) number of steps it took to reach the end node
from the start node.
}
\description{
Traverses a network using a switcher-random walk process and records the
number of steps required to get from node \code{start} to node \code{observe}.
}
\details{
If a node specified in \code{observe} has never been visited then the function
returns \code{nmax} for that node.
}
