% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randwalk.R
\name{rand_walk}
\alias{rand_walk}
\title{Random walk}
\usage{
rand_walk(g, agg = "median", nrep = 100, nmax = 1000, pjump = 0,
  verbose = FALSE)
}
\description{
Random walk simulates the average number of steps between two vertices
for each pair of vertices in the graph.
}
