% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{rename_memnet}
\alias{rename_memnet}
\title{Rename memnet objects}
\usage{
rename_memnet(dat, names)
}
\arguments{
\item{dat}{lists of indices that correspond to positions in a name vector.}

\item{names}{character vector giving the names.}
}
\value{
lists of character vectors.
}
\description{
Function replaces the index vectors created by \link{adjlist} or \link{fluency}
by their original names.
}
