% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_graph}
\alias{rw_graph}
\title{Create random walk graph}
\usage{
rw_graph(dat)
}
\arguments{
\item{dat}{}
}
\value{
A matrix
}
\description{
Create a random walk graph from verbal fluency data that includes edges for words
that occur within a window size of 1.
}
