% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{search_rw_mean}
\alias{search_rw_mean}
\title{Search network repeatedly using switcher-random walk process}
\usage{
search_rw_mean(adjlist, start, observe, nmax = 1000L, pjump = 0,
  type = 0L, nrep = 100L)
}
\arguments{
\item{adjlist}{a list containing row indices of nodes adjacent node to the ith
node as created by \link{get_adjlist}.}

\item{start}{index of the start vertix.}

\item{observe}{integer vector specifying the nodes whose first visits should be recorded.}

\item{nmax}{integer specifying the maximum number of steps.}

\item{pjump}{numeric specifying the probability of a jump.}

\item{type}{integer controlling network start and jump nodes.
For \code{type = 0} the process selects the start node and any jump
nodes proportional to their degree. For \code{type = 1} the process
selects a random node to serve both as the start node and the jump node.
For \code{type = 2} the process selects the start and any jump nodes
uniformly at random.}
}
\value{
Numeric, 3 column matrix containing in each row the start node, the end node, and
the (minimum) number of steps it took to reach the end node from the start node.
}
\description{
Traverses a network using a switcher-random walk process repeatedly, records
the earliest visit to nodes of interest and averages the result.
}
\details{
Beginning at a given start node, function traverses a network using switcher
random walk and records for each of a list of nodes of interest the index at which
the respective nodes have been visited first.

If a node specified in \code{observe} has never been visited then the function
returns \code{nmax} for that node.
}
