% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{threshold_graph}
\alias{threshold_graph}
\title{Create threshold graph}
\usage{
threshold_graph(dat, min_cooc = 1L)
}
\arguments{
\item{min_cooc}{integer specifying the minimum number of times two words
have to coocur within a window size of \code{l} to consider including
an edge between them.}
}
\value{
A matrix
}
\description{
Create a graph from verbal fluency data by adding edges for words that occur
adjacent to each other more frequently than \code{min_cooc}.
}
