% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{aspl}
\alias{aspl}
\title{Average shortest path length}
\usage{
aspl(g, type = "corrected", directed = FALSE)
}
\arguments{
\item{g}{a graph object of class \code{igraph}.}

\item{directed}{a logical specifying whether the algorithm should
take edge weights into account \code{igraph}.}
}
\value{
Local clustering coefficient.
}
\description{
Computes the average shortest path length using \code{igraph}'s
\link[igraph]{average.path.length} function.
}
\details{
Per default the corrected average shortest path length is computed
as the average shortest path divided by
\code{log(n vertices)/log(average degree)}, i.e., the expected
average shortest path length for an Erdös-Renyi random graph.
}
