% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fluency}
\alias{fluency}
\title{Verbal fluency generator}
\usage{
fluency(adjlist, n, pjump, type)
}
\arguments{
\item{adjlist}{a list containing row indices for adjacent vertices as created
by \link{get_adjlist}.}

\item{n}{integer specifying the number of productions.}

\item{pjump}{numeric specifying the probability of a jump.}

\item{type}{integer controlling network start and jump vertices.
For \code{type = 0} the process selects the start vertex and any jump
vertices proportional to their degree. For \code{type = 1} the process
selects a random vertex to serve as the start vertex and the jump vertex.
For \code{type = 2} the process selects the start and any jump vertices
uniformly.}
}
\value{
Integer vector containing the indices of the fluency productions.
  Indices refer to the row of the item in the original adjacency matrix. See
  \link{get_adjlist}.
}
\description{
Generates verbal fluency data using a switcher-random walk process.
}
\details{
Function produces verbal fluency data via a switcher random walk
process that traverses the network by selecting a neighbor with
probability 1-\code{pjump} or jumps to a random place in the network
with probability \code{pjump}. Where the random walk process enters
the network and where it jumps to is additionally controlled
by \code{type}. Neighbors are always selected uniformly.
}
