% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_adjlist}
\alias{get_adjlist}
\title{Get adjacency list}
\usage{
get_adjlist(adj)
}
\arguments{
\item{adj}{numeric matrix specifying the adjacency matrix.}
}
\value{
A list of vectors containing the indices of adjacent vertices.
}
\description{
Get list containing adjacent vertices for each vertex in the graph.
}
\details{
Adjacent vertices are returned in terms of their row index of the adjacency matrix.
}
