% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{k_dist}
\alias{k_dist}
\title{Maximum difference between cumulative degree distribution}
\usage{
k_dist(g1, g2)
}
\arguments{
\item{g1}{a graph object of class \code{igraph}.}

\item{g2}{a graph object of class \code{igraph}.}
}
\value{
Distance between cumulative degree distributions in terms
of the difference of cumulative proportions.
}
\description{
Determines the maximum difference between the cumulative
degree distributions of two graphs.
}
