% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{l_comp}
\alias{l_comp}
\title{Retrieve largest component}
\usage{
l_comp(g)
}
\arguments{
\item{g}{graph object of class \code{igraph}.}
}
\value{
A list containing the largest component as an \code{igraph}
object and the relative size of the largest component compared
to the entire graph.
}
\description{
Retrieves the largest component. In case of equally sized
components the first is component is retrieved.
}
