% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grplot.R
\name{neighborhood_plot}
\alias{neighborhood_plot}
\title{Neighborhood plot}
\usage{
neighborhood_plot(g, v, k, col = "steelblue", ...)
}
\arguments{
\item{g}{graph object of class \code{igraph}. Alternatively, a
matrix or data.frame with the first two columns specifying the
from- and to-vertices.}

\item{v}{integer specifying the row index (within the adjacency
matrix) of the vertex whose the neighborhood should be plotted.
Alternatively the vertex name, provided g allows for the extraction
of a name attribute.}

\item{k}{integer specifying the size of the neighborhood. Specifically,
the plot will contain all vertices that are \code{k} or fewer steps away
from \code{v}.}

\item{\dots}{arguments to be passed to \link{graph_plot}.}
}
\description{
Plot k-Neighborhood of given vertex containing all vertices with distances
≤ k from given vertex.
}
