% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{one_ffluency}
\alias{one_ffluency}
\title{Fast verbal fluency generator}
\usage{
one_ffluency(adj_list, n, pjump, type)
}
\arguments{
\item{adj_list}{a list containing row indices of nodes adjacent node to the ith
node as created by \link{get_adjlist}.}

\item{n}{integer specifying the maximum number of productions. Function may
return fewer than \code{n}.}

\item{pjump}{numeric specifying the probability of a jump.}

\item{type}{integer controlling network start and jump nodes.
For \code{type = 0} the process selects the start node and any jump
nodes proportional to their degree. For \code{type = 1} the process
selects a random node to serve both as the start node and the jump node.
For \code{type = 2} the process selects the start and any jump nodes
uniformly at random.}
}
\value{
Integer vector containing the indices of the fluency productions.
  Indices refer to the row of the item in the original adjacency matrix. See
  \link{get_adjlist}.
}
\description{
Generates verbal fluency data using a switcher-random walk process.
}
\details{
Function produces verbal fluency data via a switcher random walk
process that traverses the network by selecting a neighbor with
probability \code{1-pjump} or jumps to a random place in the network
with probability \code{pjump}. Where the random walk process enters
the network and where it jumps to is further controlled
by \code{type}. Neighbors are always selected uniformly.

In contrast to \link{fluency}, this function does not check at every step
whether the sampled neighbor is already in the list of productions. Instead,
\code{ffluency} simply returns the list of unique productions. This means
that if repetitions occur \code{ffluency} will produce sequences of length
\code{min(n * 3 - k, n)} where k is the number of repeptitions.
}
