% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{threshold_graph}
\alias{threshold_graph}
\title{Create threshold graph}
\usage{
threshold_graph(dat, min_cooc = 2L)
}
\arguments{
\item{dat}{list of character vectors containing the fluency productions.}

\item{min_cooc}{integer specifying the minimum number of times two words
are required to coocur one step apart from each other for an edge to
connect those words.}
}
\value{
A matrix
}
\description{
Create a graph from verbal fluency data by adding edges for words that occur
adjacent to each other more frequently than \code{min_cooc}.
}
